+++ plugins/module_utils/ksconf_shared.py
@@ -7,7 +7,9 @@ __metaclass__ = type
 import os
 import re
 
+from ansible.errors import AnsibleActionFail
 from ansible.module_utils.basic import AnsibleModule
+from ansible.utils.display import Display
 
 
 __version__ = "0.18.1"
@@ -31,18 +33,27 @@ def find_splunk_home():
     return None
 
 
-def check_ksconf_version(module):
-    # type: (AnsibleModule) -> tuple
+def check_ksconf_version(module: AnsibleModule = None) -> tuple:
+    if not module:
+        display = Display()
     try:
         from ksconf import __version__ as ksconf_version
     except ImportError:
-        module.fail_json(msg="Unable to import the 'ksconf' python module.  "
-                             "Try running 'pip install -U ksconf'")
+        message = "Unable to import the 'ksconf' python module.  "\
+                  "Try running 'pip install -U ksconf'"
+        if module:
+            module.fail_json(msg=message)
+        else:
+            raise AnsibleActionFail(message=message)
 
     match = re.match(r'(\d+)\.(\d+)\.(\d+)(.*)$', ksconf_version)
     if match:
         p = match.groups()
         return int(p[0]), int(p[1]), int(p[2]), p[3]
     else:
-        module.warn("Unable to parse ksconf version.  '{}'".format(ksconf_version))
+        message = f"Unable to parse ksconf version.  '{ksconf_version}'"
+        if module:
+            module.warn(message)
+        else:
+            display.warning(message)
         return 0, 0, 0, ksconf_version